gestao_pulverizacao_backend/requirements.txt
Django
djangorestframework
djangorestframework-gis
psycopg2-binary
rasterio
geopandas
